src/gpu: K6bc — naming + CUDA polish batch (selected items K6.22-K6.51)
common/gpu_huffman.cuh:
  K6.27 — lutTotalLen / lutNumSyms return uint8_t (was int) for type-
    consistency with packLutEntry's inputs and lutSym1/2's returns.
  K6.28 — split LUT_NUM_SYMS_* tag values from LUT_MAX_SYMS_PER_STEP
    with comment distinguishing wire-format tag vs per-step count.
  K6.29 — document lutSymPair endianness (sym1 low byte, sym2 high byte).

common/gpu_wire_format.cuh:
  K6.22 — comment explaining absence of RAW_CHUNK_TYPE = 0 (implicit
    default-zero; raw type compared inline as `ct == 0`).
  K6.24 — alias STREAM_HEADER_BYTES → LZ_SUBSTREAM_COUNT_HDR_BYTES with
    back-compat retain.
  K6.25 — rewrite the 5-byte entropy header comment with the actual
    bit layout (byte 0: type:4 | dm1_hi4:4; bytes 1..4 BE:
    dm1_low14:14 | comp_size:18) instead of "see writeHuffChunkHdr".

common/gpu_warp.cuh:
  K6.51 — add SLZ_GUARD_SINGLE_THREAD() macro and apply at 5 single-
    thread orchestration kernel entry points (walk_frame, prefix_sum,
    merge_huff_descs, compact_huff_descs, compact_raw_descs). Intent
    self-documenting at the call site.

encode/lz_format.cuh:
  K6.31 — Fibonacci-multiplier naming symmetry: HASH_MUL_FIB_K6,
    HASH_MUL_FIB_64, HASH_MUL_A as primary names; old names
    (FIB_HASH_MUL_K6, FIB_HASH_MUL_64, HASH_A_MUL) retained as
    back-compat aliases.

decode/walk_frame_kernel.cuh:
  K6.45 — walkReadF32LE: union type-pun → memcpy bit-cast (C++17
    defined-behavior idiom; nvcc emits the same single load).
  K6.51 — apply SLZ_GUARD_SINGLE_THREAD().

decode/lz_decode_general.cuh:
  K6.39 — remove the never-taken `break` at the end of the per-block
    loop body; the dead writes that would run on the last iter are
    DCE'd by nvcc. Added comment justifying the trade.

decode/prefix_sum_chunks_kernel.cuh + decode/compact_descs_kernels.cuh
+ decode/merge_huff_descs_kernel.cuh:
  K6.51 — apply SLZ_GUARD_SINGLE_THREAD().

decode/descriptors.zig:
  K6.32 — primary name `WALK_MAX_CHUNKS` (SCREAMING_SNAKE for module
    constants); back-compat `walk_max_chunks` alias retained.

encode/assemble_kernel.cu:
  K6.61 — comment the two __syncwarp() in emitEntropyStream (publish
    lane-0 header write before cooperative body copy).

REG/STACK: all kernels unchanged.
Verify: -gpu enwik8 L3 SHA matches 2b49720e... ✓

+++ src/gpu/common/gpu_huffman.cuh
@@ -100,11 +100,17 @@ static constexpr int LUT_SYM2_SHIFT = 8;
 static constexpr int LUT_LEN_SHIFT  = 16;
 static constexpr int LUT_NSYM_SHIFT = 24;
 
-// num_syms field values in a packed LUT entry.
+// num_syms field values in a packed LUT entry — these are wire-format
+// tag values stored in the LUT's high byte, not symbol counts (though
+// SINGLE=1 and DUAL=2 happen to match the count of symbols they emit).
 static constexpr uint8_t LUT_NUM_SYMS_SINGLE   = 1;   // single-symbol entry
 static constexpr uint8_t LUT_NUM_SYMS_DUAL     = 2;   // double-symbol entry
 static constexpr uint8_t LUT_NUM_SYMS_ESCAPE   = 3;   // height-limit escape entry
-static constexpr int     LUT_MAX_SYMS_PER_STEP = 2;   // max symbols an X2 entry emits
+
+// Distinct from the tag values above: this is the maximum number of
+// symbols any LUT entry can emit per decode step (an X2 entry emits 2).
+// Used to size per-thread output buffers in the decode loop.
+static constexpr int     LUT_MAX_SYMS_PER_STEP = 2;
 
 __device__ __forceinline__ uint32_t packLutEntry(uint8_t sym1, uint8_t sym2,
                                                   uint8_t total_len, uint8_t num_syms) {
@@ -115,8 +121,11 @@ __device__ __forceinline__ uint32_t packLutEntry(uint8_t sym1, uint8_t sym2,
 }
 
 // Field accessors mirroring packLutEntry — keep unpack centralized.
+// All four return uint8_t for type-consistency with packLutEntry's inputs;
+// lutSymPair returns uint16_t because callers consume it as a packed
+// (sym1 in low byte, sym2 in high byte) value, e.g. for `*(uint16_t*)dst = lutSymPair(e)`.
 __device__ __forceinline__ uint8_t  lutSym1(uint32_t entry)     { return (uint8_t)(entry >> LUT_SYM1_SHIFT); }
 __device__ __forceinline__ uint8_t  lutSym2(uint32_t entry)     { return (uint8_t)(entry >> LUT_SYM2_SHIFT); }
 __device__ __forceinline__ uint16_t lutSymPair(uint32_t entry)  { return (uint16_t)(entry & 0xFFFFu); }
-__device__ __forceinline__ int      lutTotalLen(uint32_t entry) { return (entry >> LUT_LEN_SHIFT)  & 0xFF; }
-__device__ __forceinline__ int      lutNumSyms(uint32_t entry)  { return (entry >> LUT_NSYM_SHIFT) & 0xFF; }
+__device__ __forceinline__ uint8_t  lutTotalLen(uint32_t entry) { return (uint8_t)((entry >> LUT_LEN_SHIFT)  & 0xFF); }
+__device__ __forceinline__ uint8_t  lutNumSyms(uint32_t entry)  { return (uint8_t)((entry >> LUT_NSYM_SHIFT) & 0xFF); }
